kafka/requirements.txt
kafka-python-ng==2.2.3
requests==2.32.3
hdfs==2.7.3
python-dotenv==1.0.1
